python/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
supabase==2.4.0
python-dotenv==1.0.1
pydantic==2.9.2
pytest==8.3.3
httpx==0.24.1
pytest-asyncio==0.23.8
email-validator==2.2.0
gotrue==2.7.0